fix: 메인화면 모든기록 수정
